+++ packages/sources/finnhub-secondary/test/integration/adapter-rest.test.ts
@@ -46,21 +46,12 @@ describe('rest', () => {
       const data = {
         base: 'EUR',
         overrides: {
-          finnhub: { EUR: 'OANDA:EUR_USD' },
+          'finnhub-secondary': { EUR: 'OANDA:EUR_USD' },
         },
       }
       mockResponseSuccess()
       const response = await testAdapter.request(data)
-      expect(response.statusCode).toBe(200)
-      expect(response.json()).toMatchSnapshot()
-    })
 
-    it('should return success for requests overriden by default adapter overrides', async () => {
-      const data = {
-        base: 'EUR',
-      }
-      mockResponseSuccess()
-      const response = await testAdapter.request(data)
       expect(response.statusCode).toBe(200)
       expect(response.json()).toMatchSnapshot()
     })
@@ -75,12 +66,45 @@ describe('rest', () => {
       expect(response.statusCode).toBe(200)
       expect(response.json()).toMatchSnapshot()
     })
+
+    it('should return success for standard pairs, when pair has inverse config', async () => {
+      const data = {
+        base: 'USD',
+        quote: 'JPY',
+      }
+      mockResponseSuccess()
+      const response = await testAdapter.request(data)
+      expect(response.statusCode).toBe(200)
+      expect(response.json()).toMatchSnapshot()
+    })
+
+    it('should return success for inverted pairs', async () => {
+      const data = {
+        base: 'JPY',
+        quote: 'USD',
+      }
+      mockResponseSuccess()
+      const response = await testAdapter.request(data)
+      expect(response.statusCode).toBe(200)
+      expect(response.json()).toMatchSnapshot()
+    })
   })
 
   describe('forex endpoint (quote alias)', () => {
     it('should return success', async () => {
       const data = {
         base: 'AAPL',
+      }
+      mockResponseSuccess()
+      const response = await testAdapter.request(data)
+      expect(response.statusCode).toBe(200)
+      expect(response.json()).toMatchSnapshot()
+    })
+
+    it('should return success when providing base and quote', async () => {
+      const data = {
+        base: 'AAPL',
+        quote: 'USD',
       }
       mockResponseSuccess()
       const response = await testAdapter.request(data)