+++ packages/sources/finnhub/src/index.ts
@@ -2,6 +2,7 @@ import { expose, ServerInstance } from '@chainlink/external-adapter-framework'
 import { PriceAdapter } from '@chainlink/external-adapter-framework/adapter'
 import { quote, buildQuoteEndpoint } from './endpoint'
 import { config } from './config'
+import includes from './config/includes.json'
 
 const rateLimiting = {
   tiers: {
@@ -21,6 +22,7 @@ const adapter = new PriceAdapter({
   config,
   endpoints: [quote],
   rateLimiting,
+  includes,
 })
 
 const server = (): Promise<ServerInstance | undefined> => expose(adapter)