+++ packages/composites/bsol-price/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @chainlink/bsol-price-adapter
 
+## 2.2.65
+
+### Patch Changes
+
+- Updated dependencies [011aec39e]
+  - @chainlink/token-allocation-adapter@1.11.6
+
+## 2.2.64
+
+### Patch Changes
+
+- Updated dependencies [1a00fdfc1]
+- Updated dependencies [715221438]
+  - @chainlink/token-allocation-adapter@1.11.5
+
 ## 2.2.63
 
 ### Patch Changes