+++ packages/composites/linear-finance/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @chainlink/linear-finance-adapter
 
+## 2.3.54
+
+### Patch Changes
+
+- Updated dependencies [011aec39e]
+  - @chainlink/token-allocation-adapter@1.11.6
+
+## 2.3.53
+
+### Patch Changes
+
+- Updated dependencies [1a00fdfc1]
+- Updated dependencies [715221438]
+  - @chainlink/token-allocation-adapter@1.11.5
+
 ## 2.3.52
 
 ### Patch Changes