+++ packages/composites/por-indexer/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @chainlink/por-indexer-adapter
 
+## 1.2.63
+
+### Patch Changes
+
+- Updated dependencies [011aec39e]
+  - @chainlink/proof-of-reserves-adapter@1.18.7
+
+## 1.2.62
+
+### Patch Changes
+
+- Updated dependencies [1a00fdfc1]
+- Updated dependencies [715221438]
+  - @chainlink/proof-of-reserves-adapter@1.18.6
+
 ## 1.2.61
 
 ### Patch Changes