+++ packages/composites/synth-index/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @chainlink/synth-index-adapter
 
+## 1.2.65
+
+### Patch Changes
+
+- Updated dependencies [011aec39e]
+  - @chainlink/token-allocation-adapter@1.11.6
+
+## 1.2.64
+
+### Patch Changes
+
+- Updated dependencies [1a00fdfc1]
+- Updated dependencies [715221438]
+  - @chainlink/token-allocation-adapter@1.11.5
+
 ## 1.2.63
 
 ### Patch Changes