+++ packages/sources/blocksize-capital/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @chainlink/blocksize-capital-adapter
 
+## 2.0.8
+
+### Patch Changes
+
+- 011aec39e: Bumped framework version
+
+## 2.0.7
+
+### Patch Changes
+
+- 1a00fdfc1: Bumped framework version
+- 715221438: Bumped framework version
+
 ## 2.0.6
 
 ### Patch Changes