+++ packages/sources/cfbenchmarks/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @chainlink/cfbenchmarks-adapter
 
+## 2.1.6
+
+### Patch Changes
+
+- 011aec39e: Bumped framework version
+
+## 2.1.5
+
+### Patch Changes
+
+- 1a00fdfc1: Bumped framework version
+- 715221438: Bumped framework version
+
 ## 2.1.4
 
 ### Patch Changes