+++ packages/sources/coinmarketcap/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @chainlink/coinmarketcap-adapter
 
+## 2.0.12
+
+### Patch Changes
+
+- 011aec39e: Bumped framework version
+
+## 2.0.11
+
+### Patch Changes
+
+- 1a00fdfc1: Bumped framework version
+- 715221438: Bumped framework version
+
 ## 2.0.10
 
 ### Patch Changes