+++ packages/sources/coinmetrics/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @chainlink/coinmetrics-adapter
 
+## 3.2.7
+
+### Patch Changes
+
+- 011aec39e: Bumped framework version
+
+## 3.2.6
+
+### Patch Changes
+
+- 1a00fdfc1: Bumped framework version
+- 715221438: Bumped framework version
+
 ## 3.2.5
 
 ### Patch Changes