+++ packages/sources/dxfeed-secondary/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @chainlink/dxfeed-secondary-adapter
 
+## 2.0.6
+
+### Patch Changes
+
+- 011aec39e: Bumped framework version
+- Updated dependencies [011aec39e]
+  - @chainlink/dxfeed-adapter@2.0.6
+
+## 2.0.5
+
+### Patch Changes
+
+- 1a00fdfc1: Bumped framework version
+- 715221438: Bumped framework version
+- Updated dependencies [1a00fdfc1]
+- Updated dependencies [715221438]
+  - @chainlink/dxfeed-adapter@2.0.5
+
 ## 2.0.4
 
 ### Patch Changes