+++ packages/sources/galaxy/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @chainlink/galaxy-source-adapter
 
+## 2.1.16
+
+### Patch Changes
+
+- 011aec39e: Bumped framework version
+
+## 2.1.15
+
+### Patch Changes
+
+- 1a00fdfc1: Bumped framework version
+- 715221438: Bumped framework version
+
 ## 2.1.14
 
 ### Patch Changes