+++ packages/sources/ncfx/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @chainlink/ncfx-adapter
 
+## 3.3.7
+
+### Patch Changes
+
+- 011aec39e: Bumped framework version
+
+## 3.3.6
+
+### Patch Changes
+
+- 1a00fdfc1: Bumped framework version
+- 715221438: Bumped framework version
+
 ## 3.3.5
 
 ### Patch Changes