+++ packages/sources/polkadot-balance/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @chainlink/polkadot-balance-adapter
 
+## 1.2.4
+
+### Patch Changes
+
+- 011aec39e: Bumped framework version
+
+## 1.2.3
+
+### Patch Changes
+
+- 1a00fdfc1: Bumped framework version
+- 715221438: Bumped framework version
+
 ## 1.2.2
 
 ### Patch Changes