+++ packages/sources/stader-address-list/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @chainlink/stader-address-list-adapter
 
+## 2.3.6
+
+### Patch Changes
+
+- 011aec39e: Bumped framework version
+
+## 2.3.5
+
+### Patch Changes
+
+- 1a00fdfc1: Bumped framework version
+- 715221438: Bumped framework version
+
 ## 2.3.4
 
 ### Patch Changes